services/citizen_transfer/requirements.txt
aniso8601==9.0.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
circuitbreaker==2.0.0
click==8.1.7
cryptography==43.0.1
Flask==3.0.3
Flask-RESTful==0.3.10
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pika==1.3.2
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
urllib3==2.2.3
Werkzeug==3.0.4